refactor: move query-specific extends files into extends/query/
Group the query-component files (resolveQueryHooks / runQueryHooks /
hookedOperations / wrapControllerWithHooks) under src/util/extends/query/
so future extension components (e.g. result) can live in sibling
directories. buildExtendedClient stays at extends/ as the shared client
assembly hub. Pure move + import-path fixes, no behavior change.
